Update elasticsearch to latest version
This updates elasticsearch to the latest version for mitaka work.  This
also removes the site plugins that we were using in the legacy stack
as they have been deprecated in this version of elasticsearch.  I've
removed the mappings file as it was causing errors and the new version
of elasticsearch is doing a good job of dynamically configuring the
mappings.

Connects rcbops/u-suk-dev#175