Update versions of Elasticsearch pip packages
The elasticsearch pip package for the new version of Elasticsearch that
we are using should be updated to 2.3.0.

In order to utilise 2.3.0 we have to override the mitaka default upper
constraint for OpenStack which is version 1.9.0

Additionally, the elasticsearch-curator pip package should be updated to
keep up to date, and to ensure everything remains compatible.

Connected https://github.com/rcbops/u-suk-dev/issues/262